\name{RcmdrPlugin.survival-internal}
\title{Internal RcmdrPlugin.survival Objects}
\alias{allVarsClusters}
\alias{CoxDfbeta}
\alias{CoxDfbetas}
\alias{CoxModel}
\alias{coxphP}
\alias{CoxZPH}
\alias{highOrderTermsP}
\alias{listCoxModels}
\alias{listSurvregModels}
\alias{MartingalePlots}
\alias{myTkProgressBar}
\alias{NumericOrDate}
\alias{padNA}
\alias{PartialResPlots}
\alias{PlotCoxph}
\alias{startStop}
\alias{Survdiff}\alias{powerExample}
\alias{Survfit}
\alias{survregModel}
\alias{survregP}
\alias{TermPlots}
\alias{toDate}
\description{Internal RcmdrPlugin.survival objects.}
\details{These are not to be called by the user.}
\keyword{internal}
