\name{RcmdrPlugin.survival-package}
\alias{RcmdrPlugin.survival-package}
\alias{RcmdrPlugin.survival}
\docType{package}
\title{
Rcmdr Plug-In Package for the survival Package
}
\description{
This package provides an R Commander plug-in for the survival package,
with dialogs for managing survival data (this to a limited extent),
Cox models, parametric survival regression models,
estimation of survival curves, testing for differences in survival curves,
and a variety of diagnostics, tests, and displays.
}

\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.survival\cr
Type: \tab Package\cr
Version: \tab 0.6-1\cr
Date: \tab 2009-01-27\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

The plug-in is tightly integrated with the \emph(R Commander) interface; see the following menus: 
\textbf{Data -> Survival data"}, \textbf{Statistics -> Survival analysis}, 
\textbf{Statistics -> Fit Models}, \textbf{Models -> Hypothesis tests},
\textbf{Models -> Numerical diagnostics}, \textbf{Models -> Graphs}.
}

\author{
John Fox

Maintainer: John Fox <jfox@mcmaster.ca>
}

\section{Acknowledgment}{
I am grateful to Marilia Sa Carvalho, FIOCRUZ, Rio de Janeiro, Brazil, for many comments suggestions.
}

\keyword{ package }
