\name{SurvivalData}
\alias{SurvivalData}

\title{Define Survival Data Dialog Box}
\description{
 This dialog box permits you to define a time variable (or start and stop variables),
 an event indicator, a strata variable or variables, and a cluster variable to be associated with the
 current data set. If these characteristics are defined, then they will become default
 choices where appropriate in other dialog boxes.
}

\usage{
SurvivalData() # normally not called directly
}

\value{
Used only for its side effect.
}

\author{John Fox <jfox@mcmaster.ca>}

\references{
	John Fox, Marilia Sa Carvalho (2012).
    The RcmdrPlugin.survival Package: Extending the R Commander Interface to Survival Analysis.
    \emph{Journal of Statistical Software}, 49(7), 1-32.
} 

\keyword{survival}
