\name{RcmdrPlugin.survival-package}
\alias{RcmdrPlugin.survival-package}
\alias{RcmdrPlugin.survival}
\docType{package}
\title{
Rcmdr Plug-In Package for the survival Package
}
\description{
This package provides an R Commander plug-in for the survival package,
with dialogs for Cox models, parametric survival regression models,
estimation of survival curves, and testing for differences in survival curves.
}

\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.survival\cr
Type: \tab Package\cr
Version: \tab 0.5-0\cr
Date: \tab 2008-12-14\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
John Fox

Maintainer: John Fox <jfox@mcmaster.ca>
}

\keyword{ package }
