\name{RcmdrPlugin.survival-package}
\alias{RcmdrPlugin.survival-package}
\alias{RcmdrPlugin.survival}
\docType{package}
\title{
Rcmdr Plug-In Package for the survival Package
}
\description{
This package provides an R Commander plug-in for the survival package,
with dialogs for managing survival data (this to a limited extent),
Cox models, parametric survival regression models,
estimation of survival curves, testing for differences in survival curves,
and a variety of diagnostics, tests, and displays.
}

\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.survival\cr
Type: \tab Package\cr
Version: \tab 0.7-1\cr
Date: \tab 2009-03-03\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

The plug-in is tightly integrated with the \emph{R Commander} interface; see the following menus: 
\strong{Data -> Survival data"}, \strong{Statistics -> Survival analysis}, 
\strong{Statistics -> Fit Models}, \strong{Models -> Hypothesis tests},
\strong{Models -> Numerical diagnostics}, \strong{Models -> Graphs}.
}

\author{
John Fox

Maintainer: John Fox <jfox@mcmaster.ca>
}

\section{Acknowledgments}{
I am grateful to Marilia Sa Carvalho, FIOCRUZ, Rio de Janeiro, Brazil, for many comments suggestions, 
and to the following individuals for translations of messages into other languages: Philippe Grojean
(French), Anton Korobeinikov (Russian), and Marilia Sa Carvalho (Portuguese).
}

\keyword{ package }
